fix udp code
This was changed ages ago but not updated here. The current code *here*
conflicts with SHA1 (in the multicodec table).

Ideally, we wouldn't duplicate these tables but *this* table lists *sizes*. We
could also just remove the numbers from this table and *only* list names/sizes
but we can do that later.