Start the conformance test suite
The multiaddr spec defines or will define the following things:

- Protocol numbering and values
- Human-readable and packed representations of multiaddrs
- Converting between human-readable and packed representations
- Certain operations like: encapsulate, decapsulate, filter, match, resolve

We'll want to continuously test every implementation for correctness and completeness.

It seems useful to have:

- a single source of test cases
- a multiaddr CLI tool per implementation (also useful beyond testing)
- conformance testing as part of each implementation's CI

---

As a first step, this PR adds:

- Textual test cases in cucumber/gherkin format
- A basic runtime for these test cases,
  to run against a configurable multiaddr CLI (i.e. one per language impl)
- Makefile for running the tests against the known multiaddr CLIs

The go-multiaddr CLI is added in another PR: multiformats/go-multiaddr#99.
That PR also pulls in the conformance test suite and runs it in CI.
The CLI takes a multiaddr in string or packed representation,
and prints a JSON object with detailed information about the multiaddr.

This is all obviously far from complete - nevertheless I want to get it set up
end-to-end quickly and then iterate fast.
There are more CLI impls to add (js, rust, etc.) and many more test cases to write.