Adding webrtc-aside.
This transport is yet an other webrtc transport but instead of current it doesn't use any meeting server or any protocol.
Instead it use a libp2p stream so it can use circuit or ws for SDP trickle exchange.
The only drawback is it require routing.
Then its DTLS for security and SCTP for muxing.