Define `/unix`
Adds a protocol note for how to encode paths to Unix domain sockets
as strings, that may include the delimiting character of `/`.

This allows us to append other tuples to the multiaddr while also
ensuring we can round-trip the address to a string and back.

This doesn't affect the binary representation of the multiaddr since
everything is length-delimited.

Takes inspiration from #164 and proposes using URI encoding for the
segment, the same as the `/http-path` component.

One difference is if the path is to represent the filesystem root, it
must be included in the value portion of the tuple, otherwise it can
be omitted.